Fix SSH connection leak...again.
This goes the brute force route by explicitly closing SSH
connections each time we are finished using an OVN client.

The biggest change here is that we comment out the pre-creation
of ovs-ssh clients in OvnScenario. It is the biggest source of
leaking connections, and the connections aren't actually used in
the scenario. The idea behind it is sound, but its use here doesn't
work for us.

I don't like this fix. I think we can get away with fewer SSH
connections, and I feel like we can code this in a way that doesn't
require explicit calls to close the SSH sessions. See the large
XXX comment in OvnScenario for some thoughts on the matter.